+++ lib/custom_adaptive_scaffold.dart
@@ -10,10 +10,10 @@ export "src/adaptive_layout.dart";
 export "src/adaptive_scaffold.dart";
 export "src/adaptive_scaffold_controller.dart";
 export "src/breakpoints.dart";
-export "src/destination/adaptive_scaffold_destination.dart";
+export "src/navigation_bar/destination.dart";
 export "src/navigation_bar/navigation_bar.dart";
-export "src/navigation_bar/navigation_bar_destination.dart";
 export "src/navigation_destination.dart";
+export "src/navigation_rail/destination.dart";
 export "src/navigation_rail/navigation_rail.dart";
-export "src/navigation_rail/navigation_rail_destination.dart";
+export "src/navigation_shared/adaptive_scaffold_destination.dart";
 export "src/slot_layout.dart";
