+++ lib/src/navigation_bar/destination_info.dart
