+++ lib/src/navigation_bar/theme_defaults.dart
