+++ lib/src/navigation_shared/adaptive_scaffold_destination.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import "../material.dart";
+import "../_internal_material.dart";
 import "../navigation_destination.dart";
 
 /// A convenience destination for use with [AdaptiveScaffold] that mirrors the
