+++ lib/src/navigation_shared/navigation_indicator.dart
